refactor(httpapi): remove hono shim from folder
